% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn-daa-data.R
\name{adorn_category_option_combos}
\alias{adorn_category_option_combos}
\title{Adorn Category Option Combo Metadata}
\usage{
adorn_category_option_combos(
  df,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{df}{Dataframe containing DAA data indicator data to be adorned.}

\item{d2_session}{R6 session object.}
}
\value{
df An adorned dataframe
}
\description{
Adorn Category Option Combo Metadata
}
