% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn-daa-data.R
\name{adorn_daa_data}
\alias{adorn_daa_data}
\title{Adorn DAA Indicator Data}
\usage{
adorn_daa_data(
  df,
  include_coc = FALSE,
  aggregate_indicators = TRUE,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{df}{Dataframe containing DAA data indicator data to be adorned.}

\item{include_coc}{Boolean indicating whether Category Option Combo data
should be kept or removed from returned dataset.}

\item{aggregate_indicators}{Indicates whether indicators should be rolled up
across "Age" and "Age Aggregate" data elements. Also will filter out
certain data elements from 2017 and 2018 that would cause duplication.}

\item{d2_session}{R6 session object.}
}
\value{
df An adorned dataframe.
}
\description{
Cleans and adorns a dataframe of DAA data containing UIDs with indicator
names, fiscal years, and weighted concordance and discordance information,
as well as other information.
}
