% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn-daa-data.R
\name{adorn_indicators}
\alias{adorn_indicators}
\title{Adorn Indicator Metadata}
\usage{
adorn_indicators(
  df,
  aggregate_names,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{df}{Dataframe containing DAA data indicator data to be adorned.}

\item{aggregate_names}{Indicates whether indicators should be rolled up
across "Age" and "Age Aggregate" data elements. Also will filter out
certain data elements from 2017 and 2018 that would cause duplication.}

\item{d2_session}{R6 session object.}
}
\value{
df An adorned dataframe
}
\description{
Adorn Indicator Metadata
}
