% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn-weights.R
\name{adorn_weights}
\alias{adorn_weights}
\title{Adorn DAA Indicator Data with Weighted Metrics for All Levels}
\usage{
adorn_weights(
  daa_indicator_data = NULL,
  ou_hierarchy = NULL,
  pvls_emr = NULL,
  adorn_level6 = FALSE,
  adorn_emr = FALSE
)
}
\arguments{
\item{daa_indicator_data}{Dataframe containing DAA indicator data.}

\item{ou_hierarchy}{Dataframe containing the Organisational hierarchy.}
}
\value{
A dataframe of DAA Indicator data with weightings and weighted
discordance and concordance calculated for levels 3 through 5.
}
\description{
Adorn DAA Indicator Data with Weighted Metrics for All Levels
}
