% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attribute_data}
\alias{attribute_data}
\title{Site attribute data.}
\format{
A data frame with XXXXX rows and XX variables:
\describe{
\item{item_name}{description}
\item{item_name}{description}
...
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
attribute_data
}
\description{
A dataset containing the site attribute data for DAA countries.
}
\keyword{datasets}
