% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_cache}
\alias{check_cache}
\title{Check for availability and freshness of cached dataset}
\usage{
check_cache(cache_path, max_cache_age = NULL)
}
\arguments{
\item{cache_path}{Path to the cached file to be checked.}

\item{max_cache_age}{Maximum age at which the cache is considered "fresh".}
}
\value{
cache Returns the cached dataset if it is available and
fresh, otherwise returns NULL.
}
\description{
Check for availability and freshness of cached dataset
}
