% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coc_metadata}
\alias{coc_metadata}
\title{Category Option Combo metadata.}
\format{
A data frame with 2 variables:
\describe{
\item{categoryoptioncomboid}{The numeric code used to identify
the Category Option Combination in DATIM.}
\item{categoryoptioncomboname}{The name of the Category Option
Combination in DATIM.}
...
}
}
\source{
\url{https://www.datim.org/}
}
\usage{
coc_metadata
}
\description{
A dataset containing the metadata for category option combinations.
}
\keyword{datasets}
