% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{current_fiscal_year}
\alias{current_fiscal_year}
\title{Return current FY based on system date.}
\usage{
current_fiscal_year()
}
\value{
Current FY as numeric.
}
\description{
Return current FY based on system date.
}
