% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daa.analytics-package.R
\docType{package}
\name{daa.analytics-package}
\alias{daa.analytics}
\alias{daa.analytics-package}
\title{daa.analytics: Utilities for Compiling and Analyzing PEPFAR Data Alignment Activity Data}
\description{
General functions for pulling, joining, analyzing, and visualizing DAA data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pepfar-datim/daa-analytics}
  \item Report bugs at \url{https://github.com/pepfar-datim/daa-analytics}
}

}
\author{
\strong{Maintainer}: Chris Nemarich \email{cnemarich@guidehousefederal.com}

}
\keyword{internal}
