% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daa.analytics-package.R
\name{daa_analytics_params}
\alias{daa_analytics_params}
\title{Standardized package function parameter definitions}
\usage{
daa_analytics_params(
  aggregate_indicators,
  aws_s3_bucket,
  cache_path,
  coc_metadata,
  combined_data,
  d2_session,
  daa_indicator_data,
  dataset_name,
  de_metadata,
  df,
  fiscal_year,
  geo_session,
  grouping_columns,
  include_coc,
  max_cache_age,
  ou_hierarchy,
  ou_metadata,
  ou_uid,
  pe_metadata,
  pvls_emr,
  pvls_emr_raw,
  weighting_name,
  weights_list,
  ...
)
}
\arguments{
\item{aggregate_indicators}{Indicates whether indicators should be rolled up
across "Age" and "Age Aggregate" data elements. Also will filter out
certain data elements from 2017 and 2018 that would cause duplication.}

\item{aws_s3_bucket}{The URL for the particular bucket being accessed.}

\item{cache_path}{Path to the cached file to be checked.}

\item{coc_metadata}{Dataframe containing category option combination
metadata.}

\item{combined_data}{Dataframe containing adorned DAA indicator data combined with PVLS & EMR data.}

\item{d2_session}{R6 session object.}

\item{daa_indicator_data}{Dataframe containing DAA indicator data.}

\item{dataset_name}{The name of the dataset to be returned from S3.}

\item{de_metadata}{Dataframe containing data element metadata.}

\item{df}{Dataframe containing data to be adorned.}

\item{fiscal_year}{List of fiscal years starting in October to be included.}

\item{geo_session}{R6 session object specifically for a DHIS2 GeoAlign session.}

\item{grouping_columns}{A list of columns on which weighting
groups should be based.}

\item{include_coc}{Boolean indicating whether Category Option Combo data
should be kept or removed from returned dataset.}

\item{max_cache_age}{Maximum age at which the cache is considered "fresh".}

\item{ou_hierarchy}{Dataframe containing DATIM organisation unit hierarchy.}

\item{ou_metadata}{Dataframe containing organisation unit metadata.}

\item{ou_uid}{UID for the Operating Unit whose data is being queried.}

\item{pe_metadata}{Dataframe containing period metadata.}

\item{pvls_emr}{Dataframe of PVLS and EMR data joined with metadata.}

\item{pvls_emr_raw}{Unadorned dataframe of PVLS and EMR indicator data.}

\item{weighting_name}{Name to be given to weighting column.}

\item{weights_list}{A list containing the levels at which concordance
metrics should be calculated. \code{OU} level is the default, calculating
sites as a share of the overall country. SNU levels 1-3 are supported,
as well as \code{EMR}, which groups by whether or not a facility has an EMR.}

\item{...}{Additional arguments to pass.}
}
\value{
list of all parameters of this constructor function
}
\description{
Standardized package function parameter definitions
}
\concept{parameter-helpers}
