% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daa_indicator_data}
\alias{daa_indicator_data}
\title{DAA Indicator Data.}
\format{
A data frame with XXXXX rows and XX variables:
\describe{
\item{item_name}{description}
\item{item_name}{description}
...
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
daa_indicator_data
}
\description{
A dataset containing the data for DAA indicators for all DAA countries
for each year of the activity.
}
\keyword{datasets}
