% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daa_indicators}
\alias{daa_indicators}
\title{DAA indicator list.}
\format{
A data frame with 14 rows and 3 variables:
\describe{
\item{uid}{The alphanumeric UID associated with the indicator in DATIM.}
\item{indicator}{The short name of the indicator.}
\item{notes}{Notes on years and disaggregates relevant to the DAA.}
}
}
\source{
\url{https://www.datim.org/}
}
\usage{
daa_indicators
}
\description{
A dataset containing the indicator names and IDs for the DAA.
}
\keyword{datasets}
