% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_availability}
\alias{data_availability}
\title{Data availability.}
\format{
A data frame with XXXXX rows and XX variables:
\describe{
\item{item_name}{description}
\item{item_name}{description}
...
}
}
\source{
\url{http://www.geoalign.org/}
}
\usage{
data_availability
}
\description{
A dataset containing information regarding whether a country provided
an mapping and data for DAA indicators in each year of the activity.
}
\keyword{datasets}
