% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de_metadata}
\alias{de_metadata}
\title{Data Element metadata.}
\format{
A data frame with XXXXX rows and XX variables:
\describe{
\item{item_name}{description}
\item{item_name}{description}
...
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
de_metadata
}
\description{
A dataset containing the metadata for DATIM data elements.
}
\keyword{datasets}
