% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetch_s3_files}
\alias{fetch_s3_files}
\title{Fetch data from S3 bucket.}
\usage{
fetch_s3_files(s3, Bucket, Key_1, Key_2)
}
\arguments{
\item{s3}{The s3 object created by the paws package containing
user credentials.}

\item{Key_1}{The folder extension where the DAA folders are stored.}

\item{Key_2}{The specific folder where the DAA data being accessed is stored.}

\item{bucket}{The URL for the particular bucket being accessed.}
}
\value{
A dataframe of the data located in the specified S3 sub-bucket.
}
\description{
Extracts data from a specific S3 bucket, saves it to a folder, and returns
a dataframe with the same data.
}
