% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-attributes.R
\name{get_attribute_table}
\alias{get_attribute_table}
\title{Fetch MOH ID and attributes from DATIM}
\usage{
get_attribute_table(
  ou_uid,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{ou_uid}{UID for the Operating Unit whose data is being queried}

\item{d2_session}{DHIS2 Session id for the DATIM session.}
}
\value{
A dataframe of OU site-level attributes including name, id,
MOH ID, and longitude and latitude of the site.
}
\description{
Gets data on site attributes for a specific operating unit from DATIM or
DATIM4U, including site name, id, MOH ID, longitude, and latitude.
}
