% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pvls-emr-data.R
\name{get_coc_metadata}
\alias{get_coc_metadata}
\title{Fetch Category Option Combo Metadata}
\usage{
get_coc_metadata(s3, aws_s3_bucket)
}
\arguments{
\item{s3}{The s3 object created by the paws package containing
user credentials.}

\item{aws_s3_bucket}{The URL for the particular bucket being accessed.}
}
\value{
Datatable relating Category Option Combo IDs to DATIM codes.
}
\description{
Fetches category option combo metadata from S3. Intended to be combined
with PVLS and EMR indicator data.
}
