% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-daa-countries.R
\name{get_daa_countries}
\alias{get_daa_countries}
\title{Fetch List of Participating Countries from GeoAlign}
\usage{
get_daa_countries(geo_session)
}
\arguments{
\item{geo_session}{DHIS2 Session id for the GeoAlign session.}
}
\value{
A dataframe of DAA country names, UIDs, three-letter acronyms,
and facility level.
}
\description{
Extracts list of all countries that participate in the DAA, including
name, UID, three-letter acronym, and facility level.
}
