% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-daa-data.R
\name{get_daa_data}
\alias{get_daa_data}
\title{Get DAA Indicator Data}
\usage{
get_daa_data(
  ou_uid,
  fiscal_year,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{ou_uid}{UIDs for the Operating Units whose data are being queried.}

\item{fiscal_year}{Fiscal years for which data should be gathered.}

\item{d2_session}{DHIS2 Session id for the DATIM session.}
}
\value{
Dataframe of unadorned PEPFAR and the MOH DAA indicator data.
}
\description{
Fetches DAA indicator data for both PEPFAR and the MOH partner for a single
country.
}
