% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-geoalign-data.R
\name{get_data_availability}
\alias{get_data_availability}
\title{Fetch Indicator Mapping and Data Availability from GeoAlign}
\usage{
get_data_availability(geo_session = geo_session)
}
\arguments{
\item{geo_session}{DHIS2 Session id for the GeoAlign session.}
}
\value{
A dataframe of indicator mapping, disaggregation level, and data
availability organized by activity year and country.
}
\description{
Extracts all data for all countries and activity years from GeoAlign regarding
whether countries have provided indicator mappings, the disaggregation
level, and whether data was imported for that indicator.
}
