% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-daa-data.R
\name{get_data_value_sets}
\alias{get_data_value_sets}
\title{Get Data Value Sets}
\usage{
get_data_value_sets(
  parameters,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{parameters}{Dataframe of key-value pairs that
will be taken as parameters of the API call.}

\item{d2_session}{R6 session object}
}
\value{
Dataframe with six columns.
}
\description{
Get Data Value Sets
}
