% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-daa-data.R
\name{get_dataset_uids}
\alias{get_dataset_uids}
\title{Get Data Set UIDs}
\usage{
get_dataset_uids(fiscal_year = NULL)
}
\arguments{
\item{fiscal_year}{List of fiscal years starting in October.}
}
\value{
Filtered dataframe of fiscal years and dataSets
for DAA indicator data.
}
\description{
Get Data Set UIDs
}
