% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-import-history.R
\name{get_import_history}
\alias{get_import_history}
\title{Fetch Indicator Mapping and Data Availability from GeoAlign}
\usage{
get_import_history(geo_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{geo_session}{DHIS2 Session id for the GeoAlign session.}
}
\value{
A dataframe of indicator mapping, disaggregation level, and data
availability organized by activity year and country.
}
\description{
Extracts all data for all countries and activity years from GeoAlign
regarding whether countries have provided indicator mappings, the
disaggregation level, and whether data was imported for that indicator.
}
