% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-orgunit-info.R
\name{get_org_unit_info}
\alias{get_org_unit_info}
\title{Get Organisation Unit Information from DATIM.}
\usage{
get_org_unit_info(ou_name, d2_session)
}
\arguments{
\item{ou_name}{Name for the Operating Unit whose data is being queried.}

\item{d2_session}{DHIS2 Session id for the DATIM session.}
}
\value{
A list object containing the DATIM country level, DATIM facility
level, and the ISO3 country code.
}
\description{
Fetches Organisation Unit data for a particular country, including the
organisation hierarchy levels for the country and facility, as well as the
ISO3 country code.
}
