% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-orgunit-info.R
\name{get_ou_name}
\alias{get_ou_name}
\title{Get Organisation Unit Name from UID.}
\usage{
get_ou_name(ou_uid)
}
\arguments{
\item{ou_uid}{UID for the Operating Unit whose data is being queried.}
}
\value{
A string containing the country name.
}
\description{
Returns the country name based on the organisation unit UID.
}
