% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pvls-emr-data.R
\name{get_pvls_emr_table}
\alias{get_pvls_emr_table}
\title{Fetch Raw PVLS and EMR Indicator Data.}
\usage{
get_pvls_emr_table(s3, aws_s3_bucket, last_update = NULL, folder = "data")
}
\arguments{
\item{s3}{The s3 object created by the paws package containing
user credentials.}

\item{aws_s3_bucket}{The URL for the particular bucket being accessed.}

\item{last_update}{The datetime when the most recent file was created. Used
to validate whether new data needs to be retrieved from S3. If NULL, always
retrieves dataset from S3 regardless of last update time.}

\item{folder}{The folder extension where the output file should be saved. If
missing, saves to the \code{data} folder in the working directory.}
}
\value{
Dataframe containing raw PVLS and EMR data.
}
\description{
Fetches raw PVLS and EMR indicator data from S3. This data is encoded and
must be combined with the appropriate metadata to be interpreted.
}
