% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-s3-data.R
\name{get_s3_data}
\alias{get_s3_data}
\title{Fetch DAA S3 Data set}
\usage{
get_s3_data(
  aws_s3_bucket = Sys.getenv("AWS_S3_BUCKET"),
  dataset_name = NULL,
  cache_folder = NULL
)
}
\arguments{
\item{aws_s3_bucket}{The URL for the particular bucket being accessed.}

\item{dataset_name}{The name of the dataset to be returned from S3.}
}
\value{
data A dataframe of the indicated data set.
}
\description{
Fetches the indicated DAA data set from S3, applies appropriate
naming, and returns as a dataframe. If no S3 Bucket is provided
but a cache_folder is, then will attempt to retrieve a cache file.
If no cache_folder is provided, then will not check for an existing
cache and will not save a copy of the retrieved datset to a new or
updated cache file.
}
