% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-geoalign-data.R
\name{get_upload_timestamps}
\alias{get_upload_timestamps}
\title{Fetch Import Timestamps from GeoAlign}
\usage{
get_upload_timestamps(geo_session)
}
\arguments{
\item{geo_session}{DHIS2 Session id for the GeoAlign session.}
}
\value{
A dataframe of
}
\description{
Extracts all data for all countries and activity years from GeoAlign
regarding whether a country has completed each step of the DAA process
with timestamps for completion.
}
