% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pvls-emr-data.R
\name{getDAAPEMetadata}
\alias{getDAAPEMetadata}
\title{Fetches Period Metadata}
\usage{
getDAAPEMetadata(s3, aws_s3_bucket)
}
\arguments{
\item{s3}{The s3 object created by the paws package containing
user credentials.}

\item{aws_s3_bucket}{The URL for the particular bucket being accessed.}
}
\value{
Datatable relating Period IDs to DATIM codes.
}
\description{
Fetches period metadata from S3. Intended to be combined with PVLS and EMR
indicator data.
}
