% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pvls-emr-data.R
\name{getOUMetadata}
\alias{getOUMetadata}
\title{Fetches Organsation Unit Metadata}
\usage{
getOUMetadata(s3, aws_s3_bucket)
}
\arguments{
\item{s3}{The s3 object created by the paws package containing
user credentials.}

\item{aws_s3_bucket}{The URL for the particular bucket being accessed.}
}
\value{
Datatable relating Organisation Unit IDs to DATIM codes.
}
\description{
Fetches organisation unit metadata from S3. Intended to be combined with
PVLS and EMR indicator data.
}
