% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{import_history}
\alias{import_history}
\title{Country DAA Import History.}
\format{
A data frame with 5 variables:
\describe{
\item{namelevel3}{The name of the country participating in the DAA.}
\item{period}{The reporting period for the DAA.}
\item{indicator}{The reporting indicator for the DAA.}
\item{has_disag_mapping}{The disaggregate mapping provided by the country
for the indicator in the given reporting period. Values can either be
'Fine', 'Coarse', or 'Total'. Countries that did not provide a mapping are
marked as 'No'.}
\item{has_results_data}{Indicates whether a country reported data for the
associated indicator during the most recent reporting period. Values are
recorded as 'Yes' or 'No'. Periods prior to the most recent reporting
period are recorded as 'NA'.}
}
}
\source{
\url{http://www.geoalign.org/}
}
\usage{
import_history
}
\description{
A dataset containing information regarding whether a country provided
an mapping and data for DAA indicators in each year of the activity.
}
\keyword{datasets}
