% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ou_hierarchy}
\alias{ou_hierarchy}
\title{Organisation Unit Hierarchy.}
\format{
A data frame with XXXXX rows and XX variables:
\describe{
\item{item_name}{description}
\item{item_name}{description}
...
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
ou_hierarchy
}
\description{
A dataset containing the organisation unit hierarchies of all DAA countries.
}
\keyword{datasets}
