% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ou_metadata}
\alias{ou_metadata}
\title{Organisation Unit metadata.}
\format{
A data frame with XXXXX rows and XX variables:
\describe{
\item{item_name}{description}
\item{item_name}{description}
...
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
ou_metadata
}
\description{
A dataset containing the metadata for DATIM organisation units.
}
\keyword{datasets}
