% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pe_metadata}
\alias{pe_metadata}
\title{Period metadata.}
\format{
A data frame with 2 variables:
\describe{
\item{periodid}{The numeric ID used in DATIM for the time period.}
\item{iso}{The ISO date/time value for the period.}
...
}
}
\source{
\url{https://www.datim.org/}
}
\usage{
pe_metadata
}
\description{
A dataset containing the metadata for periods.
}
\keyword{datasets}
