% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s3_datasets}
\alias{s3_datasets}
\title{S3 Datasets}
\format{
A dataframe with 3 variables:
\describe{
\item{dataset_name}{Plain english name of the dataset, as defined within
the daa.analytics package.}
\item{key}{The S3 location where the dataset can be found.}
\item{filters}{Any filters that should be applied to the data for
cleaning, if relevant.}
}
}
\usage{
s3_datasets
}
\description{
A dataset containing all the relevant metadata datasets available on S3,
their name, S3 key, and the filters that should be applied to the dataset
for cleaning prior to return.
}
\keyword{datasets}
