% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn-weights.R
\name{weighted_concordance}
\alias{weighted_concordance}
\title{Calculate Weighted Concordance.}
\usage{
weighted_concordance(df, weighting_name, grouping_columns)
}
\arguments{
\item{df}{Dataframe containing data to be adorned.}

\item{weighting_name}{Name to be given to weighting column.}

\item{grouping_columns}{A list of columns on which weighting
groups should be based.}
}
\value{
A single value for the weighted concordance of the site.
}
\description{
Calculates the weighted concordance for a given site using the total
number of patients reported by the MOH and PEPFAR as well as the
weighting factor.
}
