% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{weighted_discordance}
\alias{weighted_discordance}
\title{Calculate Weighted Discordance.}
\usage{
weighted_discordance(moh, pepfar, weighting)
}
\arguments{
\item{moh}{The number of patients reported by the MOH at the site.}

\item{pepfar}{The number of patients reported by PEPFAR at the site.}

\item{weighting}{The weighting factor given to the site.}
}
\value{
A single value for the weighted discordance of the site.
}
\description{
Calculates the weighted discordance for a given site using the total
number of patients reported by the MOH and PEPFAR as well as the
weighting factor.
}
