% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-daa-data.R
\name{weighting_levels}
\alias{weighting_levels}
\title{Adorn DAA Indicator Data with Weighted Metrics for All Levels}
\usage{
weighting_levels(daa_indicator_data = NULL, ou_hierarchy = NULL)
}
\arguments{
\item{daa_indicator_data}{Dataframe containing DAA indicator data.}

\item{ou_hierarchy}{Dataframe containing the Organisational hierarchy.}
}
\value{
A dataframe of DAA Indicator data with weightings and weighted
discordance and concordance calculated for levels 3 through 5.
}
\description{
Adorn DAA Indicator Data with Weighted Metrics for All Levels
}
