% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn-daa-data.R
\name{adorn_daa_data}
\alias{adorn_daa_data}
\title{Adorn DAA Indicator Data}
\usage{
adorn_daa_data(
  df,
  include_coc = FALSE,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{df}{Dataframe containing DAA data indicator data to be adorned.}

\item{d2_session}{R6 session object.}
}
\value{
Dataframe of DAA indicator data for both PEPFAR and the MOH as well
as both discordance and concordance metrics.
}
\description{
Cleans and adorns a dataframe of DAA data containing UIDs with indicator
names, fiscal years, and weighted concordance and discordance information,
as well as other information.
}
