% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn-daa-data.R
\name{adorn_indicators}
\alias{adorn_indicators}
\title{Adorn Indicator Metadata}
\usage{
adorn_indicators(
  df,
  aggregate_names,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{df}{Dataframe containing data to be adorned.}

\item{d2_session}{R6 session object.}
}
\value{
df An adorned dataframe
}
\description{
Adorn Indicator Metadata
}
