% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adorn-weights.R
\name{adorn_weights}
\alias{adorn_weights}
\title{Adorn DAA Indicator Data with Weighted Metrics for All Levels}
\usage{
adorn_weights(
  daa_indicator_data = NULL,
  ou_hierarchy = NULL,
  pvls_emr = NULL,
  adorn_level6 = FALSE,
  adorn_emr = FALSE
)
}
\arguments{
\item{daa_indicator_data}{Dataframe containing DAA indicator data.}

\item{ou_hierarchy}{Dataframe containing the Organisational hierarchy.}

\item{pvls_emr}{Dataframe containing TX_PVLS and EMR data. Only needed if
\code{adorn_emr} is TRUE.}

\item{adorn_level6}{Boolean indicating whether to adorn weights for
SNU level 6 for countries with facilities at level 7.}

\item{adorn_emr}{Boolean indicating whether to adorn weights based
on whether EMR is or is not present at a site.}
}
\value{
A dataframe of DAA Indicator data with weightings and weighted
concordance calculated for levels 3 through 5.
}
\description{
Adorn DAA Indicator Data with Weighted Metrics for All Levels
}
