% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{attribute_data}
\alias{attribute_data}
\title{Site attribute data.}
\format{
A data frame with 5 variables:
\describe{
\item{name}{The name of the facility.}
\item{facilityuid}{The UID used by PEPFAR to identify the facility.}
\item{MOH ID}{The UID used by the MOH to identify the facility.}
\item{longitude}{Longitude of the facility as recorded in DATIM.}
\item{latitude}{Latitude of the facility as recorded in DATIM.}
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
attribute_data
}
\description{
A dataset containing the site attribute data for DAA countries.
}
\keyword{datasets}
