% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coc_metadata}
\alias{coc_metadata}
\title{Category Option Combo metadata.}
\format{
A data frame with XXXXX rows and XX variables:
\describe{
\item{item_name}{description}
\item{item_name}{description}
...
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
coc_metadata
}
\description{
A dataset containing the metadata for category option combinations.
}
\keyword{datasets}
