% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{combined_data}
\alias{combined_data}
\title{Combined DAA Dataset.}
\format{
A data frame with XXXXX rows and XX variables:
\describe{
\item{item_name}{description}
\item{item_name}{description}
...
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
combined_data
}
\description{
A dataset containing all data for the DAA activity as well as additional
indicator data from DATIM.
}
\keyword{datasets}
