% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daa_countries}
\alias{daa_countries}
\title{DAA country list.}
\format{
A data frame with 23 rows and 4 variables:
\describe{
\item{country_name}{The name of the country.}
\item{country_uid}{The alphanumeric UID associated with the country in
DATIM.}
\item{country_code}{The three letter acronym for the country.}
\item{facility_level}{The organisation unit hierarchy level at which
facilities are located for the particular country.}
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
daa_countries
}
\description{
A dataset containing the names, abbreviations, and IDs of DAA countries.
}
\keyword{datasets}
