% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daa_countries}
\alias{daa_countries}
\title{DAA country list.}
\format{
A data frame with XXXXX rows and XX variables:
\describe{
\item{item_name}{description}
\item{item_name}{description}
...
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
daa_countries
}
\description{
A dataset containing the names, abbreviations, and IDs of DAA countries.
}
\keyword{datasets}
