% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daa_indicator_data}
\alias{daa_indicator_data}
\title{DAA Indicator Data.}
\format{
A data frame with 11 variables:
\describe{
\item{facilityuid}{The alphanumeric UID used to identify the facility in
DATIM.}
\item{indicator}{The reporting indicator for the DAA.}
\item{period}{The fiscal year reporting period.}
\item{moh}{The results provided for the given indicator by the MOH during
the associated reporting period.}
\item{pepfar}{The results provided for the given indicator by PEPFAR during
the associated reporting period.}
\item{reported_by}{A text value indicating whether results were reported
by just the MOH, just PEPFAR, or both entities at the given site for the
given indicator during the reporting period.}
\item{count_of_matched_sites}{The number of facilities in a country for the
particular indicator and reporting period for which results were reported
by both the MOH and PEPFAR.}
\item{pepfar_sum_at_matched_sites}{The total results reported by PEPFAR
at all facilities in a country for the particular indicator and reporting
period.}
\item{weighting}{The PEPFAR results at the particular facility divided by
the total results reported by PEPFAR at all facilities for the given
indicator and reporting period. This figure provides the weighting value
for concordance and discordance metrics.}
\item{weighted_discordance}{The weighted discordance between the PEPFAR
and MOH reported results at the particular facility. Can be summed across
facilities grouped by country, indicator, and period to calculate the
weighted average discordance.}
\item{weighted_concordance}{The weighted concordance between the PEPFAR
and MOH reported results at the particular facility. Can be summed across
facilities grouped by country, indicator, and period to calculate the
weighted average concordance.}
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
daa_indicator_data
}
\description{
A dataset containing the data for DAA indicators for all DAA countries
for each year of the activity.
}
\keyword{datasets}
