% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daa_indicator_data}
\alias{daa_indicator_data}
\title{DAA Indicator Data.}
\format{
A data frame with 11 variables:
\describe{
\item{facilityuid}{The alphanumeric UID used to identify the facility in
DATIM.}
\item{indicator}{The reporting indicator for the DAA.}
\item{period}{The fiscal year reporting period.}
\item{moh}{The results provided for the given indicator by the MOH during
the associated reporting period.}
\item{pepfar}{The results provided for the given indicator by PEPFAR during
the associated reporting period.}
\item{reported_by}{A text value indicating whether results were reported
by just the MOH, just PEPFAR, or both entities at the given site for the
given indicator during the reporting period.}
}
}
\source{
\url{https://www.datim.org/}
}
\usage{
daa_indicator_data
}
\description{
A dataset containing the data for DAA indicators for all DAA countries
for each year of the activity.
}
\keyword{datasets}
