% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{daa_indicators}
\alias{daa_indicators}
\title{DAA indicator list.}
\format{
A data frame with XXXXX rows and XX variables:
\describe{
\item{item_name}{description}
\item{item_name}{description}
...
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
daa_indicators
}
\description{
A dataset containing the indicator names and IDs for the DAA.
}
\keyword{datasets}
