% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{de_metadata}
\alias{de_metadata}
\title{Data Element metadata.}
\format{
A data frame with 2 variables:
\describe{
\item{dataelementid}{The numeric ID associated with the
data element in DATIM.}
\item{dataelementname}{The name of the Data Element in DATIM.}
...
}
}
\source{
\url{https://www.datim.org/}
}
\usage{
de_metadata
}
\description{
A dataset containing the metadata for DATIM data elements.
}
\keyword{datasets}
