% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-s3-data.R
\name{fetch_s3_files}
\alias{fetch_s3_files}
\title{Fetch data from S3 bucket.}
\usage{
fetch_s3_files(
  s3 = paws::s3(),
  aws_s3_bucket = Sys.getenv("AWS_S3_BUCKET"),
  key = NULL,
  folder = "data-raw",
  file_name = NULL
)
}
\arguments{
\item{s3}{The s3 object created by the paws package containing
user credentials.}

\item{aws_s3_bucket}{The URL for the particular bucket being accessed.}

\item{key}{The specific folder where the DAA data being accessed
is stored on S3.}

\item{folder}{The folder where the .csv.gz temporary file should
be locally stored.}

\item{file_name}{The file name for the .csv.gz temporary file to
be locally stored.}
}
\value{
A dataframe of the data located in the specified S3 sub-bucket.
}
\description{
Extracts data from a specific S3 bucket, saves it to a folder, and returns
a dataframe with the same data.
}
