% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-daa-data.R
\name{get_daa_data}
\alias{get_daa_data}
\title{Get DAA Indicator Data}
\usage{
get_daa_data(
  ou_uid,
  fiscal_year,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{ou_uid}{UID for the Operating Unit whose data is being queried.}

\item{fiscal_year}{List of fiscal years starting in October to be included.}

\item{d2_session}{R6 session object.}
}
\value{
Dataframe of unadorned PEPFAR and the MOH DAA indicator data.
}
\description{
Fetches DAA indicator data for both PEPFAR and the MOH partner for a single
country.
}
