% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-daa-data.R
\name{get_daa_data}
\alias{get_daa_data}
\title{Get DAA Indicator Data}
\usage{
get_daa_data(ou_uid, d2_session)
}
\arguments{
\item{ou_uid}{UID for the Operating Unit whose data is being queried.}

\item{d2_session}{DHIS2 Session id for the DATIM session.}
}
\value{
Dataframe of DAA indicator data for both PEPFAR and the MOH as well
as both discordance and concordance metrics.
}
\description{
Fetches DAA indicator data for both PEPFAR and the MOH partner for a single
country.
}
