% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pvls-emr-data.R
\name{get_ou_metadata}
\alias{get_ou_metadata}
\title{Fetches Organsation Unit Metadata}
\usage{
get_ou_metadata(s3, aws_s3_bucket, last_update = NULL, folder = "data-raw")
}
\arguments{
\item{s3}{The s3 object created by the paws package containing
user credentials.}

\item{aws_s3_bucket}{The URL for the particular bucket being accessed.}

\item{last_update}{The datetime when the most recent file was created. Used
to validate whether new data needs to be retrieved from S3. If NULL, always
retrieves dataset from S3 regardless of last update time.}

\item{folder}{The folder extension where the output file should be saved. If
missing, saves to the \code{data-raw} folder in the working directory.}
}
\value{
Datatable relating Organisation Unit IDs to DATIM codes.
}
\description{
Fetches organisation unit metadata from S3. Intended to be combined with
PVLS and EMR indicator data.
}
