% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pvls-emr-data.R
\name{get_pvls_emr_table}
\alias{get_pvls_emr_table}
\title{Fetch Raw PVLS and EMR Indicator Data.}
\usage{
get_pvls_emr_table(s3, aws_s3_bucket, last_update = NULL)
}
\arguments{
\item{s3}{The s3 object created by the paws package containing
user credentials.}

\item{aws_s3_bucket}{The URL for the particular bucket being accessed.}
}
\value{
Dataframe containing raw PVLS and EMR data.
}
\description{
Fetches raw PVLS and EMR indicator data from S3. This data is encoded and
must be combined with the appropriate metadata to be interpreted.
}
