% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-s3-data.R
\name{get_s3_data}
\alias{get_s3_data}
\title{Fetch DAA S3 Data set}
\usage{
get_s3_data(
  aws_s3_bucket = Sys.getenv("AWS_S3_BUCKET"),
  dataset_name = NULL,
  cache_folder = NULL
)
}
\arguments{
\item{aws_s3_bucket}{The URL for the particular bucket being accessed.}

\item{dataset_name}{The name of the dataset to be returned.}

\item{cache_folder}{The folder extension where cache files are located.
If not provided, will not check for an existing cache and will not save
a copy of the dataset to a new or updated cache.}
}
\value{
data A dataframe of the indicated data set.
}
\description{
Fetches the indicated DAA data set from S3, applies appropriate
naming, and returns as a dataframe.
}
