% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-s3-data.R
\name{get_s3_data}
\alias{get_s3_data}
\title{Fetch DAA S3 Data set}
\usage{
get_s3_data(s3, aws_s3_bucket, dataset_name = NULL, folder = NULL)
}
\arguments{
\item{s3}{The s3 object created by the paws package containing
user credentials.}

\item{aws_s3_bucket}{The URL for the particular bucket being accessed.}

\item{folder}{The folder extension where the output file should be saved.}

\item{file_name}{The name of the dataset to be returned.}
}
\value{
data A dataframe of the indicated data set.
}
\description{
Fetches the indicated DAA data set from S3, applies appropriate
naming, and returns as a dataframe.
}
