% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_summary.R
\name{global_summary}
\alias{global_summary}
\title{Prepare Global Summary Table of DAA Data}
\usage{
global_summary(combined_data)
}
\arguments{
\item{combined_data}{The combined dataset of DAA data for processing into a
summary file.}
}
\value{
Summary table dataframe
}
\description{
Prepare Global Summary Table of DAA Data
}
