% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ou_hierarchy}
\alias{ou_hierarchy}
\title{Organisation Unit Hierarchy.}
\format{
A data frame with 7 variables:
\describe{
\item{organisationunitid}{The numeric ID used to identify the organisation
unit in DATIM.}
\item{facilityuid}{The alphanumeric UID used to identify the facility in
DATIM.}
\item{namelevel3uid}{The UID of the parent organisation unit at hierarchy
level 3 to the given facility.}
\item{namelevel4uid}{The UID of the parent organisation unit at hierarchy
level 4 to the given facility.}
\item{namelevel5uid}{The UID of the parent organisation unit at hierarchy
level 5 to the given facility.}
\item{namelevel6uid}{For countries with their facility level at hierarchy
level 6, this will represent the UID of the facility. For countries with
their facility level at hierarchy level 7, this will represent the parent
organisation unit at hierarchy level 6 to the given facility.}
\item{namelevel7uid}{For countries with their facility level at hierarchy
level 7, this will represent the UID of the facility. For countries with
their facility level at hierarchy level 6, this value will be 'NA'.}
\item{namelevel3}{The name of the parent organisation unit at hierarchy
level 3 to the given facility.}
\item{namelevel4}{The name of the parent organisation unit at hierarchy
level 4 to the given facility.}
\item{namelevel5}{The name of the parent organisation unit at hierarchy
level 5 to the given facility.}
\item{namelevel6}{For countries with their facility level at hierarchy
level 6, this will represent the name of the facility. For countries with
their facility level at hierarchy level 7, this will represent the parent
organisation unit at hierarchy level 6 to the given facility.}
\item{namelevel7}{For countries with their facility level at hierarchy
level 7, this will represent the name of the facility. For countries with
their facility level at hierarchy level 6, this value will be 'NA'.}
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
ou_hierarchy
}
\description{
A dataset containing the organisation unit hierarchies of all DAA countries.
}
\keyword{datasets}
