% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ou_metadata}
\alias{ou_metadata}
\title{Organisation Unit metadata.}
\format{
A data frame with 4 variables:
\describe{
\item{organisationunitid}{The numeric ID associated with the organisation
unit in DATIM.}
\item{path}{The UIDs for each organisation unit hierarchy level from
Global to the specific organisation unit separated by forward slashes.}
\item{name}{The name of the organisation unit.}
\item{uid}{The alphanumeric UID associated with the organisation unit
in DATIM.}
}
}
\source{
\url{https://www.datim.org/}
}
\usage{
ou_metadata
}
\description{
A dataset containing the metadata for DATIM organisation units.
}
\keyword{datasets}
