% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pvls_emr}
\alias{pvls_emr}
\title{Viral Load Suppression and Electronic Medical Records data.}
\format{
A data frame with 9 variables:
\describe{
\item{organisationunitid}{The numeric ID used to identify the organisation
unit in DATIM.}
\item{Period}{The fiscal year reporting period.}
\item{EMR - HIV Testing Services}{Boolean value indicating whether a
facility has an electronic medical records system associated with HIV
Testing Services.}
\item{EMR - Care and Treatment}{Boolean value indicating whether a
facility has an electronic medical records system associated with HIV
care and treatment services.}
\item{EMR - ANC and/or Maternity}{Boolean value indicating whether a
facility has an electronic medical records system associated with HIV
antenatal care and/or maternity services.}
\item{EMR - EID}{Boolean value indicating whether a
facility has an electronic medical records system associated with early
infant diagnosis services.}
\item{EMR - HIV/TB}{Boolean value indicating whether a
facility has an electronic medical records system associated with
Tuberculosis treatment and testing services for HIV patients.}
\item{TX_PVLS_N}{The numerator for TX_PVLS, representing the number of
patients with suppressed viral load test results documented in a given
period.}
\item{TX_PVLS_D}{The denominator for TX_PVLS, representing the number of
patients with a viral load test result documented in a given period.}
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
pvls_emr
}
\description{
A dataset containing information on the PVLS and EMR indicators for
DAA countries.
}
\keyword{datasets}
