% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pvls_emr}
\alias{pvls_emr}
\title{Viral Load Suppression and Electronic Medical Records data.}
\format{
A data frame with 6 variables:
\describe{
\item{organisationunitid}{The numeric ID used to identify the organisation
unit in DATIM.}
\item{period}{The fiscal year reporting period.}
\item{indicator}{The reporting indicator for the DAA.}
\item{emr_at_site_for_indicator}{Boolean value indicating whether a
facility has an electronic medical records system associated with the
given indicator for that row.}
\item{tx_pvls_n}{The numerator for TX_PVLS, representing the number of
patients with suppressed viral load test results documented in a given
period.}
\item{tx_pvls_d}{The denominator for TX_PVLS, representing the number of
patients with a viral load test result documented in a given period.}
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
pvls_emr
}
\description{
A dataset containing information on the PVLS and EMR indicators for
DAA countries.
}
\keyword{datasets}
