% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pvls_emr}
\alias{pvls_emr}
\title{Viral Load Suppression and Electronic Medical Records data.}
\format{
A data frame with XXXXX rows and XX variables:
\describe{
\item{item_name}{description}
\item{item_name}{description}
...
}
}
\source{
\url{http://www.datim.org/}
}
\usage{
pvls_emr
}
\description{
A dataset containing information on the PVLS and EMR indicators for
DAA countries.
}
\keyword{datasets}
